structure vector and stratification
