+++ src/idtype/index.ts
@@ -7,7 +7,7 @@
  * Created by Samuel Gratzl on 04.08.2014.
  */
 
-export {clearSelection, EVENT_REGISTER_IDTYPE, IDTypeLike, list, listAll, persist, register, restore} from './manager'
+export {clearSelection, EVENT_REGISTER_IDTYPE, IDTypeLike, list, listAll, persist, register, restore, resolve, resolveProduct} from './manager'
 export {default as SelectAble, ISelectAble} from './ASelectAble';
 export {default as IDType} from './IDType';
 export {default as AProductSelectAble, IProductSelectAble} from './AProductSelectAble';
