+++ src/vector/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Created by sam on 26.12.2016.
+ */
+
+
+export {IVector, IVectorDataDescription} from './IVector';
