+++ src/vector/loader.ts
@@ -1,0 +1,50 @@
+/* *****************************************************************************
+ * Caleydo - Visualization for Molecular Biology - http://caleydo.org
+ * Copyright (c) The Caleydo Team. All rights reserved.
+ * Licensed under the new BSD license, available at http://caleydo.org/license
+ **************************************************************************** */
+/**
+ * Created by Samuel Gratzl on 04.08.2014.
+ */
+
+import {getAPIJSON} from '../ajax';
+import {Range, parse} from '../range';
+import {IValueType} from '../datatype';
+import {IVectorDataDescription} from './IVector';
+
+export interface IVectorLoader {
+  (desc: IVectorDataDescription): Promise<{
+    readonly rowIds: Range;
+    readonly rows: string[];
+    readonly data: IValueType[];
+  }>;
+}
+
+
+export function viaAPILoader() {
+  let _loader = undefined;
+  return (desc) => {
+    if (_loader) { //in the cache
+      return _loader;
+    }
+    return _loader = getAPIJSON('/dataset/' + desc.id).then((data) => {
+      data.rowIds = parse(data.rowIds);
+      return data;
+    });
+  };
+}
+
+export function viaDataLoader(rows: string[], rowIds: number[], data: IValueType[]) {
+  let _data = undefined;
+  return () => {
+    if (_data) { //in the cache
+      return Promise.resolve(_data);
+    }
+    _data = {
+      rowIds: parse(rowIds),
+      rows: rows,
+      data: data
+    };
+    return Promise.resolve(_data);
+  };
+}
